Update notification setup error count
https://eaflood.atlassian.net/browse/WATER-4961

The legacy code keeps track of errors by using a key on the 'events.metadata.error' field.

This change updates the batch notifications logic to update the error count in the 'events' table.

This error count is used on the reporting page to show how many errors have occurred when sending the notifications.

This change also adds a new 'UpdateEventService'. The current use case is only for updating the metadata but there may be other updates required in the future so this naming has been left as open as possible to facilitate change.