Add error logging when a call to Notify fails (#1850)
* Add error logging when a call to Notify fails

https://eaflood.atlassian.net/browse/WATER-4961

As part of getting the status updates for notifications from GOV.UK Notify we have identified an issue in how we handle errors.

For some errors we store them in the 'scheduledNotifications' to align with the legacy and for the status this is no required.

This change adds the logging to the Notify services when an error occurs in Notify. We format this error (strip out the relevant information).

Notify returns failed notification within 'response.data.status' this would be something like 'technical-failure'. This is an error when Notify has tried to send the notification, and it has failed. Not a failure in the call to Notify.