Update notifications status update to batch scheduled notifications (#1854)
* Update notifications status update to batch scheduled notifications

https://eaflood.atlassian.net/browse/WATER-4961

There was a previous assumption to use the scheduled notifications event as part of the batch process. This idea introduced some complex logic to handle counting how many request we have made to Notify.

This change updates the existing fetch logic to return an array of scheduled notifications rather than an event with scheduled notifications.

This allows the batch mechanism to function the same as it does in other places.

This change also updates the notification status update service to use the batch mechanism previously mentioned.

As we are using an upsert style query (for batch inserting) the 'water.scheduledNotifications' schema wants us to provide a 'createdAt' even though all our 'scheduledNotifications' should have a 'createdAt' date.

We should not be creating any new 'scheduledNotifications' just updating them. So we provide the 'createdAt' date to be reinserted.