Add update event error count service (#1856)
* Add update event error count service

https://eaflood.atlassian.net/browse/WATER-4961

As part of the work to get status updates from notify, we need to calculate the error count for failed scheduled notifications.

This logic follows the legacy codes approach to calculating errors, but we will do this count only when we check for status updates (if there are any).

We will get a status update from notify at least once.

This change adds the logic to iterate over an array of event id's and updates the events error count in the metadata ('event.metadata.error)'.