Update jobs notifications status update
https://eaflood.atlassian.net/browse/WATER-4961

When updating the status of a notification we can get a 'soft' error from Notify. This is an error in the 'repose.body.status' and not a http error.

When an error like this occurs we consider it an error, and we save the notification status as 'error'.

When these errors occur we want to keep the events error count up to date, this is tracked in 'event.metadata.error'.

This is done by passing in an array of events into the update event error count service. This will iterate over the events and update the error count.

This change adds a call to the update event error count service to run this error count update.

An important point to note is we need to dedupe the eventIds to ensure a performant query. Otherwise, we could have hundreds / thousands of notifications with the same event id all triggering the same update.