Fix editing nil return lines (#1865)
https://eaflood.atlassian.net/browse/WATER-4811

During testing for the above ticket, a bug was discovered on the nil returns editing journey.

The scenario was a nil return being edited. The lines haven't been initiated when you edit the nil return to no longer be a nil return. This means the /multiple-entries page errors because it is looking for lines, and there aren't any. This also means the user can't enter any readings or volumes on the check page because there are no lines.
This PR fixes this.