Add notification view and model https://eaflood.atlassian.net/browse/WATER-4961
We have previously used a 'scheduledNotifications' view for the 'water.scheduledNotifications' table.
We have decided to move to a new view called 'notifications' this is, simply, because we do not schedule notifications to be sent, we just send them.

This change adds the 'notifications' table view, with the associated model, test and test helper.