Update scheduled notifications to notifications for jobs notifications
https://eaflood.atlassian.net/browse/WATER-4961

We have previously used a 'scheduledNotifications' view for the 'water.scheduledNotifications' table. We have decided to move to a new view called 'notifications' this is, simply, because we do not schedule notifications to be sent, we just send them.

This change update the notifications logic to use this new notification model and view.

Anything previously using 'scheduledNotifications' or 'scheduledNotification' has been replaced with 'notifications' and 'notification' respectively.