Fix display null lines in view return log summary
https://eaflood.atlassian.net/browse/WATER-4861

UAT quickly highlighted an issue with our new view return log page.

The new page now contains a monthly summary of the submission lines. Currently, it generates a total per month, including for monthly. If a month has no entries, i.e. the line(s) is `null` it displays '0'.

The problem for the users is they are losing 'information'. They can't tell from the summary whether the user submitted '0' for the line(s), or didn't submit anything. If you click through to the weekly or daily detail we do display null entries as empty. But we don't in the summary, which is a real problem for monthly, as there is no detail view.

This change updates the view return log page to better handle 'null' submissions.