Update return versions check page banners
https://eaflood.atlassian.net/browse/WATER-4869

Whilst working on the return log setup journey we noticed some differences between the banners on the returns versions check page and out return logs setup check page. In the interest of keeping the service consistent this PR is updated the banners on the check page to make them the same for both.