add end date for return version inserted before existing versions (#1878)
https://eaflood.atlassian.net/browse/WATER-4972

When testing WATER-4767 the QA team encountered an issue with the return version setup engine.

The scenario being tested imagines a new licence with a start date of 2025-09-28. A new return version is then added.

Subsequently, the B&D team spotted that the start date was wrong and amended the record in NALD. The overnight import updates the record in WRLS, and then a user adds another return version, this time starting on 2025-09-01.

This PR fixes this issue and generates an end date for the return version.