Prevent requirements marked summer & quarterly submission (#1872)
* Prevent requirements marked summer & quarterly submission

https://eaflood.atlassian.net/browse/WATER-4955

Quarterly return submissions can only be approved for licences with return in a winter/all year return cycle only.
Currently, when a return requirement with a summer cycle is selected for quarterly submissions, the system does not prevent the approval of the return.
We want to prevent this combination from being approved for returns.

This PR will implement validation preventing approval of requirement for return for summer cycle and quarterly submission.