Return logs - add a note bug (#1892)
https://eaflood.atlassian.net/browse/WATER-4807

An issue has been found when a note is entered that generates a validation error, the note text is not being returned to the user.

This PR will fix that issue by ensuring that any text that has been entered is returned to the view when a validation error occurs.