Return logs - fix weekly/monthly line label https://eaflood.atlassian.net/browse/WATER-4813
When editing/creating a monthly or weekly return log the label on each line currently just shows an end date for the week/month e.g. 30 April 2023.

This needs to be corrected so that when the return is weekly the label would show "Week ending 30 April 2023" & when monthly the day would be dropped so that it just displays "April 2023".

This PR will correct these labels when editing a return logs reading/volumes.