Update notification setup to use journey specific data from session.
https://eaflood.atlassian.net/browse/WATER-4979

We have identified a few areas of improvement to the notification setup journey.

We want to make the notifications setup check page as generic as possible.

This change removes the concept of a 'journey' from creating an event (notice). This means the creation of an event is driven by the initial session setup.

This simplifies both the code and the tests. The test no longer need to check each possible journey, but the existence of the name and subtype. This removes unnecessary mapping.