Rename some /notifications to /notices
https://eaflood.atlassian.net/browse/WATER-4781

We've been migrating and updating the legacy returns invitations and reminders journeys. Adding support for quarterly returns was the driver but this is also encompassed by our ongoing replacement of the legacy service.

What we've built so far has been done on the `/notifications` path, with folders and files all using the name 'notifications'. But we've realised this has driven us into a corner.

For example, when creating a 'returns invitation', you'll select the period and the service will determine which licences should receive a notification, and who the recipient is for each one. It'll then clean up and de-dupe the recipients, and for each one left send a 'notification'.

Behind the scenes we create an `event` record of type `notification` with a unique reference that links all the 'notifications' sent, plus the 'when' and 'who'.

> The fact its an `event` record rather than its own thing was a decision of the previous team we have to live with for now until we are generating all notifications.

We are also rebuilding the page that displays these `event` records (of type `notification`), and when you click on one we then need to display all the notifications linked to it. The problem is we've also been naming these pages and their supporting services 'notifications'. We've also rebuilt the page that displays the content of the notification and named all its parts 'notification'.

You can see the corner we've now found ourselves in. We're calling the 'thing' that links together notifications 'notifications', which means things are now confusing, or even worse, clashing.

You currently access the page that lists the `events` of type `notification` from a link labelled 'Notices' (the legacy page is known as the 'Notices report').

`notice` is as good a name as any for the 'thing' that links the notifications sent so we intend to use it going forward.

This change is about renaming anything that relates to a `notice` rather than a `notification` before to get us out of the corner we're now in!