Add a feature toggle to the water abstraction alert page 'create alert' link (#1902)
* Add a feature toggle to the water abstraction alert page 'create alert' link

https://eaflood.atlassian.net/browse/WATER-4979

We are migrating the legacy water abstraction alert pages into this repo.

This change adds a new feature toggle 'ENABLE_MONITORING_STATIONS_ALERT_NOTIFICATIONS'. This will be used to hide the 'new' journey as it goes through our development process.

This change also adds another 'NOTIFICATION_TYPES' called ' 'abstraction-alert'. This key will be used to reuse existing parts of our notification setup journey.

We do need to use the monotoning station id in the upcoming journey. We will add this id to the session with the basic data we need from the database but this will be done in another change as it feels too much for this change. But we are highlighting it as we have added it to the query string.