Refactor CheckReturnCycleService
After double checking we realised that the one place that calls `CheckReturnCycleService` _never_ passes in a 'changeDate'. With this being the case, there is no requirement to have it as an argument.

We then spotted that we were determining the start and end date for the return cycle twice: first to find the match and then when generating the data for the new return cycle. Instead, we could pass these as arguments and avoid the duplicated effort.

With that tweak, we spotted we could combine generating the data needed and creating the return cycle into a single 'thing'. Though we're removing the distinct steps, there is value in seeing the data being generated at the point of insert (it works for this because both the size and complexity of generating the data is small).

Other than that, we've expanded the JSDoc's to give a bit more context and explanation.