Update notification setup to use journey specific data from session. (#1904)
* Add a feature toggle to the water abstraction alert page 'create alert' link

https://eaflood.atlassian.net/browse/WATER-4979

We are migrating the legacy water abstraction alert pages into this repo.

This change adds a new feature toggle 'ENABLE_MONITORING_STATIONS_ALERT_NOTIFICATIONS'. This will be used to hide the 'new' journey as it goes through our development process.

This change also adds another 'NOTIFICATION_TYPES' called ' 'abstraction-alert'. This key will be used to reuse existing parts of our notification setup journey.

We do need to use the monotoning station id in the upcoming journey. We will add this id to the session with the basic data we need from the database but this will be done in another change as it feels too much for this change. But we are highlighting it as we have added it to the query string.

* chore: pre pr checks

* Update notification setup to use journey specific data from session.

https://eaflood.atlassian.net/browse/WATER-4979

We have identified a few areas of improvement to the notification setup journey.

We want to make the notifications setup check page as generic as possible.

This change removes the concept of a 'journey' from creating an event (notice). This means the creation of an event is driven by the initial session setup.

This simplifies both the code and the tests. The test no longer need to check each possible journey, but the existence of the name and subtype. This removes unnecessary mapping.