Check current return cycle exists (#1887)
https://eaflood.atlassian.net/browse/WATER-4542

As part of the testing for the quarterly returns it was discovered that we lost the code for creating a return log if one doesn't already exist along the way with the refactors.

This PR refactors the code to check for the existence of the current return cycles when doing the twice annual return log job as well as in the individual licence return log job. If things go as planned the new return cycles should be created as part of the annual jobs but in case those jobs are not ran we need the coverage in the individual licence return logs code to make sure we're don't miss any return logs.