Update notice setup ad-hoc and returns period folder structure
https://eaflood.atlassian.net/browse/WATER-5012

We are adding abstraction alerts to the notification setup journey.

This will be a more complex journey with a some additional pages.

To segregate the different 'journeys' which can lead to the 'shared' logic we have grouped joinery specific service/ presenters into their own folder.

This change should make it clear where the shared logic is, and the journey specific services and presenter are.

We are aiming to create one cohesive notification solution that multiple journeys can feed into.