Seed historic return cycles for tests (#1913)
https://eaflood.atlassian.net/browse/WATER-5007

As part of our work to replace NALD as the main source of managing returns we need to be able to handle not only new returns but historic changes to returns. This PR updates the seed function we use to populate return cycles to include the historic ones so when we're testing we have match for what is on the deployed environments.