Add abstraction alerts monitoring stations service (#1908)
Add abstraction alerts monitoring stations service

https://eaflood.atlassian.net/browse/WATER-5012

We need to use the monitoring station id to get the information needed to present in the notification setup abstraction alerts journey.

This change adds the Monitoring Station Service to fetch and format the data required for the notification setup abstraction alerts journey.

When the monitoring station id is passed as an arg to the initiate session service we will add the additional data to the session.

The query has been taken from the legacy and updated to use the the 'views'. Some fields have been omitted as we do not see a need for them currently, but they can be added back in as we build the journey.

The additional data will be used across the journey and so a fixture has been added for all the tests to use so we have a single representation of the expected initial session data.