Enter readings - safe number error https://eaflood.atlassian.net/browse/WATER-4870
During QA it was spotted that when an unsafe number is entered for the reading, the default error message displayed e.g. `"2024-04-01T00:00:00.000Z" must be a safe number`.

Whilst not in the acceptance criteria for this ticket, so technically not a bug. In this PR we will make this error message a bit easier to understand for the user.