Correctly handle nil return
Initially, we were setting the `nilReturn` flag in the session then using that to determine if we're persisting a nil return. However the rest of the return log code doesn't use this in the session -- we instead check if `session.journey` is `nil-return` to determine if the user is submitting a nil return. We therefore remove this from the session and adjust `SubmitCheckService` to check `session.journey` to determine whether we're passing `true` or `false` for `nilReturn` when creating the new submission.