Add notices setup abstraction alerts - alert type submit
https://eaflood.atlassian.net/browse/WATER-5014

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This change adds the alert-type submit service and validator.

It also adds the route and controller to for the submit alert types.