Add notices setup abstraction alert - alert threshold scaffolding
https://eaflood.atlassian.net/browse/WATER-5026

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This change adds the alert-threshold service, submit service validator, view and presenter.

It also adds the route and controller to for the alert threshold services.