Update notices setup abstraction alert - alert threshold
https://eaflood.atlassian.net/browse/WATER-5026

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This change updates the alert-threshold service, submit service validator, view and presenter to capture the users selected threshold and store it in session.

We need to use the monitoring station data to show the thresholds in the ui.

We have introduces a unique key per licence monitoring station to simplify how we interact with the session data. We can use this key to know what to render and use in the upcoming journey pages.