Add notifications setup abstraction thresholds view
https://eaflood.atlassian.net/browse/WATER-5026

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This change updates the alert-threshold submit service, view and validator. The selected thresholds will be added to the session.