Add notifications setup abstraction thresholds view (#1938)
https://eaflood.atlassian.net/browse/WATER-5026

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This change updates the alert-threshold service, view and presenter.

This change also introduces an id on the licence monitoring stations. When we have added the selected thresholds to the session we need a simple way to finding them. We can not generate a unique from the data as it would be exactly the same key in some cases and using an uuid would be necessary.