Replace Math random calls with 'safe' alternate
We mainly use `Math.random()` in our unit tests to support generating references for various types of records. Recently, for the first time we used it in our `production` logic to [generate the reference for new notices](https://github.com/DEFRA/water-abstraction-system/pull/1682).

None of this are security based, but every time we make a change that touches these areas we get a security ping from both GitHub and SonarQube.

This is because `Math.random()` is not cryptographically secure. It only returns a pseudo-random number. It is common to generate random numbers in security, for example, generating 'salt' for encryption. But if the number is not truly random, then its possible your encryption could be cracked.

That's why any use of `Math.random` sets alarm bells ringing.

So, we're safe in how we use it. But wouldn't it be nice to not have to repeatedly say so?!

This change adds a suggested alternate to `Math.random()` that is based on [Crypto](https://developer.mozilla.org/en-US/docs/Web/API/Crypto) and which will make everyone calm down!