Move randomInteger to app general lib file
If we need to generate a random integer in both `/test` and `/app` then we should move the existing function to the `/app` folder and the `app/lib/general.lib.js` we've created for these types of functions.

We also update the name to match the conventions used within `general.lib.js`.