Add notifications setup abstraction alerts check licence matches
https://eaflood.atlassian.net/browse/WATER-5026

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This changes adds the check licence matches scaffolded logic. It adds a basic controller, view, presenter and service to view the page.

The implementation for the page will be added in a later change.