Update notifications setup abstraction alerts check licence matches view
https://eaflood.atlassian.net/browse/WATER-5026

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This changes updates the check licence matches controller, view, presenter and service to render the licences previously selected on the alert thresholds page, with the same information as the legacy page.

When multiple thresholds for a licence are present they are grouped by licence, but show their own abstraction period, flow and level restrictions and the last alert type.

The remove / exclude logic will be added in another change.