Update notifications setup abstraction thresholds view
https://eaflood.atlassian.net/browse/WATER-5026

We are replacing the legacy notification abstraction alert journey. We are going through page by page and adding roughly similar page for some parts of the journey. With the intention of reusing the existing send notification logic we built previously.

This change updates a previously wrong assumption of showing all thresholds for all alert type. This is not the case.

There are two restriction types 'stop' and 'reduce'. When either of these are selected as the alert type then we need to only show those restrictions, when the alert type is something else then show all the thresholds.