Replace Math random calls with 'safe' alternate (#1941)
We mainly use `Math.random()` in our unit tests to support generating references for various types of records. Recently, we used it in our `production` logic for the first time to [generate the reference for new notices](https://github.com/DEFRA/water-abstraction-system/pull/1682).

None of this is security-based, but every time we make a change that touches these areas, we get a security ping from both GitHub and SonarQube.

This is because `Math.random()` is not cryptographically secure. It only returns a pseudo-random number. It is common to generate random numbers in security, for example, generating 'salt' for encryption. But your encryption could be cracked if the number is not truly random.

That's why any use of `Math.random` sets alarm bells ringing.

We're safe in how we use it, but wouldn't it be nice not to have to repeatedly say so?!

This change adds a suggested alternate to `Math.random()` based on [Crypto randomInt()](https://nodejs.org/api/crypto.html#cryptorandomintmin-max-callback), which will calm everyone down!