Fix displaying 0 qty lines for return logs
https://eaflood.atlassian.net/browse/WATER-5035

In [Fix display null lines in view return log summary](https://github.com/DEFRA/water-abstraction-system/pull/1881) we dealt with a situation where return submission lines with a `NULL` quantity were being shown as `0` in the UI.

On the face of it, they may appear to mean the same thing, but they are, in fact, different.

`0` means the customer has explicitly told us that no water was abstracted at that time. `NULL` means no value has been submitted.

They've now highlighted a similar issue. If the quantity is `0`, we're displaying nothing. This gets interpreted as `NULL`, i.e. the customer did not submit anything which is incorrect.

This change fixes the issue.