Update return version start data validation (#1952)
https://eaflood.atlassian.net/browse/WATER-5031

When creating a return version you will create return logs which use the existing return cycles. We found an issue with missing returns cycles and have now fixed that but it only goes back to the oldest licence we have in the system.

We are 99.999% certain no one will add a new licence with a start date before 1959-04-01. But just in case, for that 0.001% chance, we will extend the start date validation in the return version setup journey.

The validator already ensures that the start date entered equals or exceeds the licence’s start date. But should we encounter a licence with a start date before 1959-04-01 we need to stop it from being created as no return logs will be added.