Update notifications setup abstraction alerts thresholds to use groups
https://eaflood.atlassian.net/browse/WATER-5026

There was a previous assumption that we could update the view and render all the thresholds as individual items. This is not the case, and we need to 'group' thresholds by three factors - measureType, thresholdValue and thresholdUnit.

This change updates the existing logic to render only one instance of each threshold. As we will need to use the users selection in later pages we have created a 'thresholdGroup' when the licence monitoring station are formatted for the session.

This gives us a single consistent key to use across the journey, the key can be used as the value in the html, and captured in the submit service, which is added to an array (of user selected thresholds) without any transformation needed.

The legacy code used a group of multiple selection buttons causing multiple errors to render. A tweak has been made to page to allow us to use a single checkbox group and only have the single error displayed for the group. This also make out implementation on the submit much simpler.