Update notifications setup abstraction alerts fetch / determine session data (#1957)
* Update notifications setup abstraction alerts fetch / determine session data

https://eaflood.atlassian.net/browse/WATER-5026

There was a previous assumption that we needed to do a raw query to get the licence monitoring stations. This is not the case as the table and data we need already exist in the monitoring station code.

This change updates the fetch licence mounting station to use a similar fetch to the one that already exists (with some columns removed).

This change also updates the determine licence monitoring stations service to use the newly fetched data.