Add notifications setup abstraction alerts check licences table
https://eaflood.atlassian.net/browse/WATER-5027

The legacy 'abstraction-alerts/check-licence-matches' page has a table displaying the selected thresholds.

We already have a table for the monitoring stations page, we can use this table in multiple places.

This change adds / updates the logic to display the monitoring restrictions for the check licence matches page.

This change does not implement the remove link logic but link and text have been implemented to prove the marco renders correctly.

We have introduced a new macro 'licenceMonitoringStationRestrictionsTable' this will be used in multiple places and has been adapted from the original table.

A base presenter should be introduced in later changed. For now, a copy has been taken from the monitoring view presenter and used in the check licence matches presenter. The tests are also copied to achieve the same expected output. When the refactor to a common presenter is done these test can be removed as they will be covered by the base presenter.