Refactor monitoring station table
https://eaflood.atlassian.net/browse/WATER-5027

We have a requirement to show the motioning station restrictions on the notices setup alert journey.

This change adds a new macro table for monitoring station restrictions, it also creates a new monitoring station base presenter to fulfil this macro.

The existing logic has been tweaked to work with this shared component.

The table conditionally renders the action column. To facilitate this we the calling function needs to provide these action per licence monitoring station. We have no way of knowing the link or text required within the base presenter itself.