Create `CreateReturnSubmissionService` (#1959)
https://eaflood.atlassian.net/browse/WATER-4847

Broken out of PR https://github.com/DEFRA/water-abstraction-system/pull/1815

This PR implements `CreateReturnSubmissionService `. The service creates a new return submission in the db for the given return log ID and using the data passed to it.

The version number will be 1 if no previous submission exists for this return log ID; otherwise, the version number will be incremented by 1. If a previous submission does exist then it will be marked as superseded by setting its `current` property to `false`.