Update FetchWaterUndertakers service
First we remove the limit we put in. Including it means we don't get any return versions back, unless you limit the main query with a debug `whereIn()` and set a specific licence ID.

This intending to be a one off job, we're not gonna waste time digging into why!

The rest of the changes are to get all the data we need to copy the existing return version.