Update the process service
This orchestrates the new services. But the key change is it can take advantage of a transaction to ensure we are only persisting the new return version if _everything_ succeeds.

When the error occurred, the record that caused it was left in a partial state. We've run this through locally with no issues but should we get one in `production`, not only will we get a meaningful error message now, we won't leave the licence in a corrupt state.