Pass in licence end date when creating return logs (#1973)
https://eaflood.atlassian.net/browse/WATER-4547

As part of our prep to go live with quarterly returns the annual return log job was ran on pre-prod with the latest data. As part of the testing our testers saw an issue we had thought we had fixed earlier. Turns out we had only resolved the issue for individual licences when they were having their return versions changed. This PR adds the fix to the annual job by making sure the licence end date is being passed into the create return logs service.