Update notifications setup abstraction alerts thresholds
https://eaflood.atlassian.net/browse/WATER-5026

When a user changes the alert type we should reset the previously selected threshold. If a thresholds is relevant to another type of alert type then those check boxes could be selected from a different alert type.

This change clears the previously selected alert thresholds if the user change te alert type.